refactor dto support helpers
